refactor(tests): separate client protocol integration
